src/misc/requirements.txt
python-chess==0.31.4
numpy==1.19.1